requirements.txt
asyncpg
fastapi
pre-commit
psycopg2-binary
pydantic
pytest
requests
sqlalchemy
uvicorn
